dicom rpm

+++ build-pkg.sh
@@ -3,13 +3,19 @@
 # Localizations
 lrelease *.ts
 
-# Remove non-free code
-for f in $(grep -l WITH_DICOM *.h *.cpp);
-  do unifdef -o $f -UWITH_DICOM $f;
+for arg in "$@"; do
+  if $arg == 'dicom'; then dicom=1; fi
 done
-rm -fr dicom*
-sed -i '$d' beryllium.pro
-sed -i 's/beryllium/beryllium-free/g' docs/*
+
+# Remove non-free code
+if ! $dicom; then  
+  for f in $(grep -l WITH_DICOM *.h *.cpp)
+    do unifdef -o $f -UWITH_DICOM $f
+  done
+  rm -fr dicom*
+  sed -i '$d' beryllium.pro
+  sed -i 's/beryllium/beryllium-free/g' docs/* debian/* beryllium.spec
+fi
 
 distro=$(lsb_release -is)
 case $distro in
@@ -17,7 +23,7 @@ Ubuntu | Debian)  echo "Building DEB package"
     cp docs/* debian/ && dpkg-buildpackage -I.svn -I*.sh -rfakeroot 
     ;;
 "openSUSE project" | fedora)  echo "Building RPM package"
-    tar czf beryllium.tar.gz ../beryllium --exclude=.svn --exclude=*.sh && rpmbuild -ta beryllium.tar.gz
+    tar czf beryllium.tar.gz ../beryllium --exclude=.svn --exclude=*.sh && ../rpmbuild -Ddicom=$dicom -ta ../beryllium.tar.gz
     ;;
 *) echo "$distro is not supported yet"
    ;;
